He comentado el endpoint para contear secuencias del StudentSequenceController. Las secuencias no se desasignan correctamente

+++ src/main/java/com/sacristan/api/interfaces/admin/controllers/student/StudentUtilsController.java
@@ -27,7 +27,7 @@ public class StudentUtilsController {
             summary = "Get Student Sequence Count",
             description = "Get the number of sequences assigned to a specific student"
     )
-    public ResponseEntity<Integer> getSequenceCount(
+    public ResponseEntity<Long> getSequenceCount(
             @PathVariable Long id
     ) {
         return ResponseEntity.ok(service.getSequenceCount(id));
